requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
dependency-injector